feat(datasets): toggleable joint-trace isolation in the episode viewer
The joint-position chart plots every joint at once, which gets hard to
read once a dataset has more than a few (bimanual doubles the count).
Legend entries are now buttons: click one to isolate it (everything
else dims, in both the chart lines and the hover tooltip); click a
second to isolate both; click a selected one again to drop it back
out. No selection means show everything, same as before.

Extracts the chart + legend + hover-tooltip block out of EpisodeViewer
into its own JointPositionChart component — it was already fully
self-contained (hoverT/jointRanges used nowhere else in that 400+ line
component), and the new selection state fit more naturally as a third
concern there than bolted onto EpisodeViewer directly. Selection is
local state, so it persists across episode navigation within one
viewing session and resets when the dialog is closed and reopened.
